Implemented GPIO ports C3 and C4(GPIO_function.py)
Added new parameter to class GPIO_function class. "sw_control" determines if we define and switch on or off ports C3 and C4 for GPIO board.
WARNING: These ports are need only for Hawaii device. sw_control must be set to 'int(1)' in order to define and switch on or off new ports. By default, the class not define the ports, a new masters program is needed.