GPIO_function update:
New I/O ports given a strict rule of never being allowed to have both driven high at any point in time. Moved some lines around to make sure this never occurs if the "safety" process stated by FEM desinger is not followed.

ALso, thread for sync clock is now not made if the 'GPIO_function' class inputs its 'sync' parameter as zero. This is to make sure clokc is not inadveratly turned on if syn is never needed.